Fix WaveFileWriter writing zeros for 32-bit PCM samples (#651)
WriteSample(float) cast the normalised [-1,1] float to an Int32 before
scaling by UInt16.MaxValue, so every value except +/-1.0 truncated to 0
and the resulting 32-bit PCM WAV was near-silent. Scale first, then
cast: (Int32)(Int32.MaxValue * sample), mirroring the 24-bit branch.

Added a round-trip regression test that writes known float samples to a
32-bit PCM WaveFileWriter and verifies the values read back.

+++ src/NAudio.Core/Wave/WaveOutputs/WaveFileWriter.cs
@@ -405,7 +405,7 @@ public class WaveFileWriter : Stream
         }
         else if (WaveFormat.BitsPerSample == 32 && WaveFormat.Encoding == WaveFormatEncoding.Extensible)
         {
-            writer.Write(UInt16.MaxValue * (Int32)sample);
+            writer.Write((Int32)(Int32.MaxValue * sample));
             dataChunkSize += 4;
         }
         else if (WaveFormat.Encoding == WaveFormatEncoding.IeeeFloat)
